Server/requirements.txt
django==1.10
django-crispy-forms==1.6.0
django-pipeline==1.6.8
decorator==4.0.10
raven==5.0.0
celery==3.1.23
djangorestframework==3.5.1
django-rest-swagger==2.1.2
redis==2.10.6
SQLAlchemy==1.1.1
psycopg2==2.7.4
unidecode==1.0.22
